+++ crates/api/src/parser/term/reducible.rs
@@ -4,7 +4,7 @@ use num_traits::One;
 
 use crate::{parser::ucum_symbol::UcumSymbol, reducible::Reducible};
 
-use super::Term;
+use super::{Exponent, Term};
 
 impl Reducible<f64> for Term {
     fn reduce_value(&self, value: f64) -> f64 {
@@ -41,7 +41,7 @@ fn combine_term_values(
     calculated_atom: f64,
     calculated_prefix: f64,
     factor: Option<u32>,
-    exponent: Option<i32>,
+    exponent: Option<Exponent>,
 ) -> f64 {
     let a_p_product = calculated_atom * calculated_prefix;