+++ crates/api/src/parser/terms/mapper/digits.rs
@@ -1,8 +1,8 @@
 use super::{Error, Visit};
-use crate::parser::terms::term_parser::Rule as TermRule;
+use crate::parser::{term::Exponent, terms::term_parser::Rule as TermRule};
 use pest::iterators::Pair;
 
-pub(super) type Digits = i32;
+pub(super) type Digits = Exponent;
 
 impl Visit<TermRule> for Digits {
     fn visit(pair: Pair<'_, TermRule>) -> Result<Self, Error> {